@@ -7,7 +7,7 @@ Gem::Specification.new do |s|
   s.name           = 'partitioned'
   s.version        = Partitioned::VERSION
   s.license        = 'New BSD License'
-  s.date           = '2013-10-10'
+  s.date           = '2014-03-26'
   s.summary        = "Postgres table partitioning support for ActiveRecord."
   s.description    = "A gem providing support for table partitioning in ActiveRecord. Support is available for postgres and AWS RedShift databases. Other features include child table management (creation and deletion) and bulk data creating and updating."
   s.authors        = ["Keith Gabryelski", "Aleksandr Dembskiy", "Edward Slavich"]
@@ -22,6 +22,6 @@ Gem::Specification.new do |s|
   s.add_dependency 'pg'
   s.add_dependency 'bulk_data_methods'
   s.add_dependency 'activerecord-redshift-adapter'
-  s.add_development_dependency 'rails', '>= 3.2.8'
+  s.add_development_dependency 'rails', '>= 4.0.4'
   s.add_development_dependency 'rspec-rails'
 end
