@@ -1,7 +1,12 @@
 # Be sure to restart your server when you modify this file.
 
-# Your secret key for verifying the integrity of signed cookies.
+# Your secret key is used for verifying the integrity of signed cookies.
 # If you change this key, all old signed cookies will become invalid!
+
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Dummy::Application.config.secret_token = '6a4eb752a007445816effccfe01a47cc30e5cfa283867193b78a35fe14f66cff5ad3ea7768d9e7453e9182ae6f5a004cd156bec62544eb22ce6ba20186b7301a'
+# You can use `rake secret` to generate a secure secret key.
+
+# Make sure your secret_key_base is kept private
+# if you're sharing your code publicly.
+Dummy::Application.config.secret_key_base = 'ce8024abf0487c65ec9abac33f9ff022ccdde0d87f059d4336bcd804612aa6d8a56cc7b88ea6bd8b819f4e92bfe88d067dba2926c1c66ec407529b0b7caa13b1'
