@@ -24,7 +24,7 @@ module Partitioned
         let!(:default_reader) { Partitioned::PartitionedBase::Configurator::Reader.new(Employee) }
         let!(:reader) do
           reader = Partitioned::PartitionedBase::Configurator::Reader.new(Employee)
-          reader.stub!(:configurators).and_return([dsl])
+          reader.stub(:configurators).and_return([dsl])
           reader
         end
 
@@ -344,7 +344,7 @@ module Partitioned
 
             it "returns employees_partitions.p42" do
               dsl.table_name("employees_partitions.p42")
-              reader.table_name.should == "employees_partitions.p42"
+              reader.table_name(57).should == "employees_partitions.p42"
             end
 
           end # when table_name value is set by value
