@@ -4,7 +4,7 @@ module TablesSpecHelper
 
   class Company < ActiveRecord::Base
     extend BulkMethodsMixin
-    has_many :employees, :class_name => 'Company', :conditions => "companies.id = employees.companies_id"
+    has_many :employees, -> { where("companies.id == employees.companies_id") }, :class_name => 'Company'
   end
 
   def create_tables
